requirements.txt
opencv-python
tensorflow==2.12.0
flask
flask-login
flask-openid
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
flask-whooshalchemy
flask-wtf
flask-babel
guess_language
flipflop
coverage
pillow
torch==1.7.0+cu117
torchvision==0.8.1+cu117
torchaudio==0.7.0